+++ Samples/tvOS-Swift/tvOS-Swift/AppDelegate.swift
@@ -15,8 +15,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
             options.sessionTrackingIntervalMillis = 5_000
             // Sampling 100% - In Production you probably want to adjust this
             options.tracesSampleRate = 1.0
-            options.enableFileIOTracing = true
-            options.enableUserInteractionTracing = true
             options.enableAppHangTracking = true
         }