+++ Sources/Sentry/SentrySDK.m
@@ -375,16 +375,18 @@ static NSUInteger startInvocations;
  */
 + (void)close
 {
+    SENTRY_LOG_DEBUG(@"Starting to close SDK.");
     // pop the hub and unset
     SentryHub *hub = SentrySDK.currentHub;
 
-    // uninstall all the integrations
+    // Uninstall all the integrations
     for (NSObject<SentryIntegrationProtocol> *integration in hub.installedIntegrations) {
         if ([integration respondsToSelector:@selector(uninstall)]) {
             [integration uninstall];
         }
     }
     [hub removeAllIntegrations];
+    SENTRY_LOG_DEBUG(@"Uninstalled all integrations.");
 
 #if SENTRY_HAS_UIKIT
     // force the AppStateManager to unsubscribe, see