+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+### Fixes 
+
+- Ensure the current GPU frame rate is always reported for concurrent transaction profiling metrics (#2929)
+
+## 8.5.0
+
 ### Features
 
 - feat: Core data operation in the main thread (#2879)
@@ -11,6 +17,8 @@
 - Crash when serializing invalid objects (#2858)
 - Possible crash in Core Data tracking (#2865)
 - Don't send screenshots with either width or height of 0 (#2876))
+- Don't send screenshots with either width or height of 0 (#2876)
+- GPU frame alignment with stack traces in profiles (#2856)
 
 ## 8.4.0